#169 - Extracted Manifests from Repo
