+++ src/main/java/com/artipie/docker/asto/AstoRepo.java
@@ -24,26 +24,16 @@
 
 package com.artipie.docker.asto;
 
-import com.artipie.asto.Content;
-import com.artipie.asto.Key;
 import com.artipie.asto.Storage;
 import com.artipie.docker.BlobStore;
-import com.artipie.docker.Digest;
 import com.artipie.docker.Layers;
+import com.artipie.docker.Manifests;
 import com.artipie.docker.Repo;
 import com.artipie.docker.RepoName;
 import com.artipie.docker.Upload;
-import com.artipie.docker.manifest.JsonManifest;
-import com.artipie.docker.manifest.Layer;
-import com.artipie.docker.manifest.Manifest;
-import com.artipie.docker.misc.ByteBufPublisher;
-import com.artipie.docker.ref.ManifestRef;
-import java.nio.charset.StandardCharsets;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
-import java.util.stream.Stream;
 
 /**
  * Asto implementation of {@link Repo}.
@@ -87,41 +77,8 @@ public final class AstoRepo implements Repo {
     }
 
     @Override
-    public CompletionStage<Manifest> addManifest(final ManifestRef ref, final Content content) {
-        return new ByteBufPublisher(content).bytes()
-            .thenCompose(bytes -> this.blobs.put(new Content.From(bytes), new Digest.Sha256(bytes)))
-            .thenCompose(
-                blob -> {
-                    final Digest digest = blob.digest();
-                    return blob.content()
-                        .thenApply(source -> new JsonManifest(digest, source))
-                        .thenCompose(
-                            manifest -> this.validate(manifest)
-                                .thenCompose(nothing -> this.addManifestLinks(ref, digest))
-                                .thenApply(nothing -> manifest)
-                        );
-                }
-            );
-    }
-
-    @Override
-    public CompletionStage<Optional<Manifest>> manifest(final ManifestRef ref) {
-        return this.readLink(ref).thenCompose(
-            digestOpt -> digestOpt.map(
-                digest -> this.blobs.blob(digest)
-                    .thenCompose(
-                        blobOpt -> blobOpt
-                            .map(
-                                blob -> blob.content()
-                                    .<Manifest>thenApply(
-                                        source -> new JsonManifest(blob.digest(), source)
-                                    )
-                                    .thenApply(Optional::of)
-                            )
-                            .orElseGet(() -> CompletableFuture.completedFuture(Optional.empty()))
-                    )
-            ).orElseGet(() -> CompletableFuture.completedFuture(Optional.empty()))
-        );
+    public Manifests manifests() {
+        return new AstoManifests(this.asto, this.blobs, this.name);
     }
 
     @Override
@@ -143,107 +100,6 @@ public final class AstoRepo implements Repo {
                 }
                 return upload;
             }
-        );
-    }
-
-    /**
-     * Validates manifest by checking all referenced blobs exist.
-     *
-     * @param manifest Manifest.
-     * @return Validation completion.
-     */
-    private CompletionStage<Void> validate(final Manifest manifest) {
-        return manifest.config()
-            .thenCompose(
-                config -> manifest.layers().thenApply(
-                    layers -> Stream.concat(
-                        Stream.of(config),
-                        layers.stream().filter(layer -> layer.urls().isEmpty()).map(Layer::digest)
-                    )
-                )
-            )
-            .thenCompose(
-                digests -> CompletableFuture.allOf(
-                    digests.map(
-                        digest -> this.blobs.blob(digest).thenCompose(
-                            opt -> {
-                                if (opt.isEmpty()) {
-                                    throw new IllegalArgumentException(
-                                        String.format("Blob does not exist: %s", digest)
-                                    );
-                                }
-                                return CompletableFuture.allOf();
-                            }
-                        ).toCompletableFuture()
-                    ).toArray(CompletableFuture[]::new)
-                )
-            );
-    }
-
-    /**
-     * Adds links to manifest blob by reference and by digest.
-     *
-     * @param ref Manifest reference.
-     * @param digest Blob digest.
-     * @return Signal that links are added.
-     */
-    private CompletableFuture<Void> addManifestLinks(final ManifestRef ref, final Digest digest) {
-        return CompletableFuture.allOf(
-            this.addLink(new ManifestRef.FromDigest(digest), digest),
-            this.addLink(ref, digest)
-        );
-    }
-
-    /**
-     * Puts link to blob to manifest reference path.
-     *
-     * @param ref Manifest reference.
-     * @param digest Blob digest.
-     * @return Link key.
-     */
-    private CompletableFuture<Void> addLink(final ManifestRef ref, final Digest digest) {
-        return this.asto.save(
-            this.link(ref),
-            new Content.From(digest.string().getBytes(StandardCharsets.US_ASCII))
-        ).toCompletableFuture();
-    }
-
-    /**
-     * Reads link to blob by manifest reference.
-     *
-     * @param ref Manifest reference.
-     * @return Blob digest, empty if no link found.
-     */
-    private CompletableFuture<Optional<Digest>> readLink(final ManifestRef ref) {
-        final Key key = this.link(ref);
-        return this.asto.exists(key).thenCompose(
-            exists -> {
-                final CompletionStage<Optional<Digest>> stage;
-                if (exists) {
-                    stage = this.asto.value(key)
-                        .thenCompose(
-                            pub -> new ByteBufPublisher(pub).asciiString()
-                        )
-                        .<Digest>thenApply(Digest.FromString::new)
-                        .thenApply(Optional::of);
-                } else {
-                    stage = CompletableFuture.completedFuture(Optional.empty());
-                }
-                return stage;
-            }
-        );
-    }
-
-    /**
-     * Create link key from manifest reference.
-     *
-     * @param ref Manifest reference.
-     * @return Link key.
-     */
-    private Key link(final ManifestRef ref) {
-        return new Key.From(
-            RegistryRoot.V2, "repositories", this.name.value(),
-            "_manifests", ref.link().string()
         );
     }
 }
