+++ src/test/java/com/artipie/docker/asto/AstoManifestsTest.java
@@ -29,7 +29,6 @@ import com.artipie.asto.Storage;
 import com.artipie.docker.Blob;
 import com.artipie.docker.Digest;
 import com.artipie.docker.ExampleStorage;
-import com.artipie.docker.Repo;
 import com.artipie.docker.RepoName;
 import com.artipie.docker.Tag;
 import com.artipie.docker.manifest.Manifest;
@@ -44,13 +43,13 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 /**
- * Integration tests for {@link AstoRepo}.
+ * Tests for {@link AstoManifests}.
  *
- * @since 0.1
+ * @since 0.3
  * @checkstyle ClassDataAbstractionCouplingCheck (500 lines)
  */
 @SuppressWarnings("PMD.AvoidDuplicateLiterals")
-final class AstoRepoITCase {
+final class AstoManifestsTest {
 
     /**
      * Storage used in tests.
@@ -58,14 +57,14 @@ final class AstoRepoITCase {
     private Storage storage;
 
     /**
-     * Repository being tested.
+     * Repository manifests being tested.
      */
-    private Repo repo;
+    private AstoManifests manifests;
 
     @BeforeEach
     void setUp() {
         this.storage = new ExampleStorage();
-        this.repo = new AstoRepo(
+        this.manifests = new AstoManifests(
             this.storage,
             new AstoBlobs(this.storage),
             new RepoName.Simple("my-alpine")
@@ -82,7 +81,7 @@ final class AstoRepoITCase {
 
     @Test
     void shouldReadNoManifestIfAbsent() throws Exception {
-        final Optional<Manifest> manifest = this.repo.manifest(
+        final Optional<Manifest> manifest = this.manifests.get(
             new ManifestRef.FromTag(new Tag.Valid("2"))
         ).toCompletableFuture().get();
         MatcherAssert.assertThat(manifest.isPresent(), new IsEqual<>(false));
@@ -115,7 +114,7 @@ final class AstoRepoITCase {
             )
             .build().toString().getBytes();
         final ManifestRef ref = new ManifestRef.FromTag(new Tag.Valid("some-tag"));
-        final Manifest manifest = this.repo.addManifest(ref, new Content.From(data))
+        final Manifest manifest = this.manifests.put(ref, new Content.From(data))
             .toCompletableFuture().join();
         MatcherAssert.assertThat(this.manifest(ref), new IsEqual<>(data));
         MatcherAssert.assertThat(
@@ -125,7 +124,7 @@ final class AstoRepoITCase {
     }
 
     private byte[] manifest(final ManifestRef ref) {
-        return this.repo.manifest(ref)
+        return this.manifests.get(ref)
             .thenApply(
                 opt ->
                     opt.map(
